move: test helper into core-device-id
